feat: steel thread for FIFO queues
[#186768689](https://www.pivotaltracker.com/story/show/186768689)

Now that we are exposing fields that can cause the destruction
of the "aws_sqs_queue resource" we must guarantee that it won't
happen accidentally.

Using `prevent_destroy` in lifecycle is a battle-testes solution
that has proved to work flawlessly. Sadly, it requires a small
change to be introduced in our fork of the community repo.

It's a pity. But such small and localised change won't cause
any maintenance issues for merging changes from upstream.
Other approaches have been analysed but are less safe.

We may harness some of them together with `prevent_destroy` in
future commits since they can protect us against other issues
such as, csb-core's prohibit_update not working in plan updates
[#185302704](https://www.pivotaltracker.com/story/show/185302704)